backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
mysql-connector-python
pymysql
numpy
pandas
scikit-learn
tensorflow
opencv-python-headless
google-generativeai
flask-jwt-extended
flask-bcrypt
python-dotenv
email_validator

